+++ core/src/test/scala/com/github/cerst/autorequire/Readme.scala
@@ -9,7 +9,7 @@ package com.github.cerst.autorequire
 object Readme {
 
   final case class Person(age: Int, name: String) {
-    autoRequire[Person](age >= 14 && name.nonEmpty)
+    autoRequire[Person](!(age < 13) && name.nonEmpty)
   }
 
   def main(args: Array[String]): Unit = {